Day 7:Authentication  completed

+++ hospital_project/core/models.py
@@ -4,7 +4,7 @@ from django.db import models
 
 class Patient(models.Model):
     name=models.CharField(max_length=100)
-    age=models.IntegerField()
+    age=models.PositiveSmallIntegerField()
     disease=models.CharField(max_length=100)
 
     def __str__(self):
